+++ config/facebook-ads.php
@@ -1,0 +1,12 @@
+<?php
+
+return [
+
+    'app_id' => env('FB_APP_ID', ''),
+
+    'app_secret' => env('FB_APP_SECRET', ''),
+
+    'access_token' => env('FB_ACCESS_TOKEN', ''),
+
+    'refresh_token' => env('FB_REFRESH_TOKEN', '')
+];